Fixed duplicated code. Fixes #138
