+++ project/backend/project-manager/src/main/java/com/meta/project/entity/BoardList.java
@@ -7,9 +7,7 @@ import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
 import lombok.EqualsAndHashCode;
-import org.hibernate.annotations.GenericGenerator;
 
-import java.time.LocalDateTime;
 import java.util.HashSet;
 import java.util.Set;
 
@@ -18,14 +16,7 @@ import java.util.Set;
 @ToString(exclude = {"board", "cards"})
 @EqualsAndHashCode(onlyExplicitlyIncluded = true)
 @Entity
-public class BoardList {
-
-    @Id
-    @GeneratedValue(generator = "uuid")
-    @GenericGenerator(name = "uuid", strategy = "org.hibernate.id.UUIDGenerator")
-    @Column(nullable = false, updatable = false)
-    @EqualsAndHashCode.Include
-    private String id;
+public class BoardList extends BaseEntity {
 
     private String title;
 
@@ -40,18 +31,4 @@ public class BoardList {
     @OneToMany(mappedBy = "boardList", cascade = CascadeType.ALL, orphanRemoval = true)
     @JsonManagedReference("list-cards")
     private Set<Card> cards = new HashSet<>();
-
-    private LocalDateTime createdAt;
-    private LocalDateTime updatedAt;
-
-    @PrePersist
-    protected void onCreate() {
-        createdAt = LocalDateTime.now();
-        updatedAt = LocalDateTime.now();
-    }
-
-    @PreUpdate
-    protected void onUpdate() {
-        updatedAt = LocalDateTime.now();
-    }
 }
