+++ project/backend/project-manager/src/main/java/com/meta/project/entity/Comment.java
@@ -3,18 +3,11 @@ package com.meta.project.entity;
 import com.fasterxml.jackson.annotation.JsonBackReference;
 import jakarta.persistence.*;
 import lombok.Data;
-import org.hibernate.annotations.GenericGenerator;
-
-import java.time.LocalDateTime;
 
 @Data
 @Entity
 @Table(name = "comments")
-public class Comment {
-    @Id
-    @GeneratedValue(generator = "uuid")
-    @GenericGenerator(name = "uuid", strategy = "org.hibernate.id.UUIDGenerator")
-    private String id;
+public class Comment extends BaseEntity {
 
     private String text;
     private String image;
@@ -24,18 +17,4 @@ public class Comment {
     @JoinColumn(name = "card_id")
     @JsonBackReference("card-comments")
     private Card card;
-
-    private LocalDateTime createdAt;
-    private LocalDateTime updatedAt;
-
-    @PrePersist
-    protected void onCreate() {
-        createdAt = LocalDateTime.now();
-        updatedAt = LocalDateTime.now();
-    }
-
-    @PreUpdate
-    protected void onUpdate() {
-        updatedAt = LocalDateTime.now();
-    }
 }
