+++ project/backend/project-manager/src/main/java/com/meta/project/entity/Todo.java
@@ -3,18 +3,11 @@ package com.meta.project.entity;
 import com.fasterxml.jackson.annotation.JsonBackReference;
 import jakarta.persistence.*;
 import lombok.Data;
-import org.hibernate.annotations.GenericGenerator;
-
-import java.time.LocalDateTime;
 
 @Data
 @Entity
 @Table(name = "todos")
-public class Todo {
-    @Id
-    @GeneratedValue(generator = "uuid")
-    @GenericGenerator(name = "uuid", strategy = "org.hibernate.id.UUIDGenerator")
-    private String id;
+public class Todo extends BaseEntity {
 
     private String content;
     private Boolean completed = false;
@@ -23,18 +16,4 @@ public class Todo {
     @JoinColumn(name = "card_id")
     @JsonBackReference("card-todos")
     private Card card;
-
-    private LocalDateTime createdAt;
-    private LocalDateTime updatedAt;
-
-    @PrePersist
-    protected void onCreate() {
-        createdAt = LocalDateTime.now();
-        updatedAt = LocalDateTime.now();
-    }
-
-    @PreUpdate
-    protected void onUpdate() {
-        updatedAt = LocalDateTime.now();
-    }
 }
